FitSuiteManager -> FitSessionActivity (files renamed)
