+++ GUI/coregui/Views/FitWidgets/FitSessionActivity.h
